Move helper functions in OptimizerRules.h and cpp files to Optimizer.h and cpp; Remove OptimizerRules.cpp
