+++ arangod/Aql/CMakeLists.txt
@@ -70,7 +70,6 @@ add_library(arango_aql STATIC
   NonConstExpression.cpp
   NonConstExpressionContainer.cpp
   Optimizer.cpp
-  OptimizerRules.cpp
   OptimizerRulesFeature.cpp
   OptimizerUtils.cpp
   OutputAqlItemRow.cpp
